Convert IdDataControls to TypeScript
